requirements.txt
git+https://github.com/openai/whisper.git
langdetect
pytube
pdf2image
pytesseract
pillow
openai-whisper